fix: move determine_num_steps, determine_freq
