+++ tests/test_datasets/test_utils.py
@@ -8,6 +8,8 @@ from etna.datasets import TSDataset
 from etna.datasets import duplicate_data
 from etna.datasets import generate_ar_df
 from etna.datasets.utils import _TorchDataset
+from etna.datasets.utils import determine_freq
+from etna.datasets.utils import determine_num_steps
 from etna.datasets.utils import get_level_dataframe
 from etna.datasets.utils import get_target_with_quantiles
 from etna.datasets.utils import inverse_transform_target_components
@@ -348,3 +350,111 @@ def test_inverse_transform_target_components(
         inverse_transformed_target_df=inverse_transformed_target_df,
     )
     pd.testing.assert_frame_equal(obtained_inverse_transformed_components_df, inverse_transformed_components_df)
+
+
+@pytest.mark.parametrize(
+    "start_timestamp, end_timestamp, freq, answer",
+    [
+        (pd.Timestamp("2020-01-01"), pd.Timestamp("2020-01-02"), "D", 1),
+        (pd.Timestamp("2020-01-01"), pd.Timestamp("2020-01-11"), "D", 10),
+        (pd.Timestamp("2020-01-01"), pd.Timestamp("2020-01-01"), "D", 0),
+        (pd.Timestamp("2020-01-05"), pd.Timestamp("2020-01-19"), "W-SUN", 2),
+        (pd.Timestamp("2020-01-01"), pd.Timestamp("2020-01-15"), pd.offsets.Week(), 2),
+        (pd.Timestamp("2020-01-31"), pd.Timestamp("2021-02-28"), "M", 13),
+        (pd.Timestamp("2020-01-01"), pd.Timestamp("2021-06-01"), "MS", 17),
+        (0, 0, None, 0),
+        (0, 5, None, 5),
+        (3, 10, None, 7),
+    ],
+)
+def test_determine_num_steps_ok(start_timestamp, end_timestamp, freq, answer):
+    result = determine_num_steps(start_timestamp=start_timestamp, end_timestamp=end_timestamp, freq=freq)
+    assert result == answer
+
+
+@pytest.mark.parametrize(
+    "start_timestamp, end_timestamp, freq",
+    [
+        (pd.Timestamp("2020-01-02"), pd.Timestamp("2020-01-01"), "D"),
+        (5, 2, None),
+    ],
+)
+def test_determine_num_steps_fail_wrong_order(start_timestamp, end_timestamp, freq):
+    with pytest.raises(ValueError, match="Start train timestamp should be less or equal than end timestamp"):
+        _ = determine_num_steps(start_timestamp=start_timestamp, end_timestamp=end_timestamp, freq=freq)
+
+
+@pytest.mark.parametrize(
+    "start_timestamp, end_timestamp, freq",
+    [
+        (pd.Timestamp("2020-01-02"), pd.Timestamp("2020-06-01"), "M"),
+        (pd.Timestamp("2020-01-02"), pd.Timestamp("2020-06-01"), "MS"),
+        (2.2, 5, None),
+    ],
+)
+def test_determine_num_steps_fail_wrong_start(start_timestamp, end_timestamp, freq):
+    with pytest.raises(ValueError, match="Start timestamp isn't correct according to given frequency"):
+        _ = determine_num_steps(start_timestamp=start_timestamp, end_timestamp=end_timestamp, freq=freq)
+
+
+@pytest.mark.parametrize(
+    "start_timestamp, end_timestamp, freq",
+    [
+        (2, 5.5, None),
+    ],
+)
+def test_determine_num_steps_fail_wrong_start(start_timestamp, end_timestamp, freq):
+    with pytest.raises(ValueError, match="End timestamp isn't correct according to given frequency"):
+        _ = determine_num_steps(start_timestamp=start_timestamp, end_timestamp=end_timestamp, freq=freq)
+
+
+@pytest.mark.parametrize(
+    "start_timestamp, end_timestamp, freq",
+    [
+        (pd.Timestamp("2020-01-31"), pd.Timestamp("2020-06-05"), "M"),
+        (pd.Timestamp("2020-01-01"), pd.Timestamp("2020-06-05"), "MS"),
+    ],
+)
+def test_determine_num_steps_fail_wrong_end(start_timestamp, end_timestamp, freq):
+    with pytest.raises(ValueError, match="End timestamp isn't reachable with freq"):
+        _ = determine_num_steps(start_timestamp=start_timestamp, end_timestamp=end_timestamp, freq=freq)
+
+
+@pytest.mark.parametrize(
+    "timestamps,answer",
+    (
+        (pd.date_range(start="2020-01-01", periods=3, freq="M"), "M"),
+        (pd.date_range(start="2020-01-01", periods=3, freq="W"), "W-SUN"),
+        (pd.date_range(start="2020-01-01", periods=3, freq="D"), "D"),
+        (pd.Series(np.arange(10)), None),
+        (pd.Series(np.arange(5, 15)), None),
+        (pd.Series(np.arange(1)), None),
+    ),
+)
+def test_determine_freq(timestamps, answer):
+    assert determine_freq(timestamps=timestamps) == answer
+
+
+@pytest.mark.parametrize(
+    "timestamps",
+    (
+        pd.to_datetime(pd.Series(["2020-02-01", "2020-02-15", "2021-02-15"])),
+        pd.to_datetime(pd.Series(["2020-02-15", "2020-01-22", "2020-01-23"])),
+    ),
+)
+def test_determine_freq_fail_cant_determine(timestamps):
+    with pytest.raises(ValueError, match="Can't determine frequency of a given dataframe"):
+        _ = determine_freq(timestamps=timestamps)
+
+
+@pytest.mark.parametrize(
+    "timestamps",
+    (
+        pd.Series([5, 4, 3]),
+        pd.Series([4, 5, 3]),
+        pd.Series([3, 4, 6]),
+    ),
+)
+def test_determine_freq_fail_int_gaps(timestamps):
+    with pytest.raises(ValueError, match="Integer timestamp isn't ordered and doesn't contain all the values"):
+        _ = determine_freq(timestamps=timestamps)
